// src/app/shared/models/equitment-interface.ts
export class EquitmentInteface {
    id?: number;
    description?: string;
    comunication?: number;
    portCom?: number;
    baudrate?: number;
    byteSize?: number;
    parity?: string;
    dtr?: string;
    rts?: string;
    stopBits?: number;
    qrydsr?: string;
    trazarCom?: string;
    trazarPort?: string;
    handshake?: string;
    hostSocket?: string;
    hostIp?: string;
    hostIpPort?: number;
    inputFolder?: string;
    outputFolder?: string;
    domain?: string;
    username?: string;
    password?: string;
    validation?: string;
    status?: boolean;
}